requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.3
altgraph==0.17.4
annotated-types==0.7.0
antiorm==1.2.1
anyio==4.6.2.post1
arrow==1.3.0
asgiref==3.8.1
attrs==23.2.0
auto-py-to-exe==2.44.2
bcrypt==4.1.3
beautifulsoup4==4.12.3
bidict==0.23.1
binaryornot==0.4.4
blinker==1.8.2
boto3==1.34.126
botocore==1.34.126
bottle==0.13.2
bottle-websocket==0.2.9
briefcase==0.3.20
build==1.2.2.post1
certifi==2024.2.2
cffi==1.17.1
chardet==5.2.0
charset-normalizer==2.0.12
click==8.1.7
clr_loader==0.2.7.post0
colorama==0.4.6
contextlib2==21.6.0
cookiecutter==2.6.0
cryptography==42.0.8
Cython==3.0.11
DateTime==5.5
db==0.1.1
Deprecated==1.2.14
Django==5.1.2
docker==7.1.0
docutils==0.21.2
Eel==0.17.0
fastapi==0.115.2
ffmpeg==1.4
ffmpeg-python==0.2.0
filetype==1.2.0
flash==1.0.3
Flask==2.2.3
Flask-BasicAuth==0.2.0
Flask-Dance==7.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
future==1.0.0
gevent==24.10.3
gevent-websocket==0.10.1
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.59.1
greenlet==3.1.1
grpcio==1.64.1
gunicorn==22.0.0
h11==0.14.0
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
jaeger-client==4.3.0
Jinja2==3.1.4
jmespath==1.0.1
jsonify==0.5
Kivy==2.3.1
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
Kivy-Garden==0.1.5
kivy-uix==1.0.0
kivy_deps.sdl2==0.8.0
kivy_webview==0.0.0a0
logging==0.4.9.6
lxml==5.3.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
MouseInfo==0.1.3
multidict==6.0.5
mysql-connector-python==8.0.28
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opentelemetry-api==1.25.0
opentelemetry-distro==0.46b0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-exporter-otlp==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-flask==0.46b0
opentelemetry-instrumentation-wsgi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
opentracing==2.4.0
opentracing-instrumentation==3.2.1
packaging==24.1
paramiko==3.4.0
pefile==2023.2.7
Pillow==9.4.0
platformdirs==4.3.6
pluggy==1.5.0
protobuf==4.25.3
proxy_tools==0.1.0
psutil==5.9.8
psycopg2==2.9.10
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyGetWindow==0.0.9
Pygments==2.18.0
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PyJWT==2.8.0
PyMsgBox==1.0.9
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==3.2.0
pyperclip==1.8.2
pypiwin32==223
pyproject_hooks==1.2.0
PyRect==0.2.0
PyScreeze==0.1.30
pytesseract==0.3.10
pytest==8.3.4
python-dateutil==2.9.0.post0
python-docx==0.8.11
python-dotenv==1.0.1
python-engineio==4.9.1
python-slugify==8.0.4
python-socketio==5.11.3
pythonnet==3.0.5
pytweening==1.2.0
pytz==2024.2
pywebview==5.3.2
pywhatkit==5.4
pywin32==308
pywin32-ctypes==0.2.3
PyYAML==6.0.2
redis==5.0.4
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
s3transfer==0.10.1
schedule==1.1.0
simple-websocket==1.0.0
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.36
sqlparse==0.5.1
starlette==0.40.0
tesseract==0.1.3
text-unidecode==1.3
threadloop==1.0.2
thrift==0.20.0
toga-core==0.4.8
toga-winforms==0.4.8
tomli_w==1.1.0
tornado==6.4.1
travertino==0.3.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.18
URLObject==2.4.3
uvicorn==0.32.0
Werkzeug==2.2.3
whichcraft==0.6.1
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.19.2
zope.event==5.0
zope.interface==7.1.1
